suppress logging on stdout during unit test
